Part B Complete. App is fully functional

+++ src/main/java/com/example/demo/repository/MoodRepository.java
@@ -4,6 +4,6 @@ import com.example.demo.entity.Mood;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface MoodRepository extends JpaRepository<Mood, Integer> {
+public interface MoodRepository extends JpaRepository<Mood, Long> { // Changed Integer to Long
     Optional<Mood> findByName(String name);
 }
